Offline working
and it's kind of blazing fast!
Just can't compute pca if theres nothing on the context. Have to code a workaround : 
If there is less than a certain power, reject the context for pca.
This will permit to compute all events in set because all_ts is gonna be less big.